chore: rearranged files and code refactoring
